Refactor the basic java project and enhanced with code
